bugfix: Fix NPEs from BugSense
https://www.bugsense.com/dashboard/project/ad73db70#error/101833340
https://www.bugsense.com/dashboard/project/ad73db70#error/101132438
https://www.bugsense.com/dashboard/project/ad73db70#error/100675473
https://www.bugsense.com/dashboard/project/ad73db70#error/102012673
https://www.bugsense.com/dashboard/project/ad73db70#error/102126205

+++ src/paulscode/android/mupen64plusae/util/FileUtil.java
@@ -145,11 +145,14 @@ public class FileUtil
         if( folder.isDirectory() )
         {
             String[] children = folder.list();
-            for( String child : children )
+            if( children != null )
             {
-                boolean success = deleteFolder( new File( folder, child ) );
-                if( !success )
-                    return false;
+                for( String child : children )
+                {
+                    boolean success = deleteFolder( new File( folder, child ) );
+                    if( !success )
+                        return false;
+                }
             }
         }
